backend/requirements.txt
Flask
flask-cors
openai-whisper
pydub
noisereduce
librosa
soundfile